+++ trunk/src/gmt_init.h
@@ -78,7 +78,7 @@ EXTERN_MSC void GMT_hash_init (struct GMT_CTRL *GMT, struct GMT_HASH *hashnode ,
 EXTERN_MSC int GMT_getdefpath (struct GMT_CTRL *GMT, char get, char **path);
 EXTERN_MSC void GMT_extract_label (struct GMT_CTRL *GMT, char *line, char *label, struct GMT_OGR_SEG *G);
 EXTERN_MSC void GMT_check_lattice (struct GMT_CTRL *GMT, double *inc, unsigned int *registration, bool *active);
-EXTERN_MSC int GMT_check_binary_io (struct GMT_CTRL *GMT, unsigned int n_req);
+EXTERN_MSC int GMT_check_binary_io (struct GMT_CTRL *GMT, uint64_t n_req);
 EXTERN_MSC char * GMT_putparameter (struct GMT_CTRL *GMT, char *keyword);
 EXTERN_MSC void GMT_set_pad (struct GMT_CTRL *GMT, unsigned int npad);
 EXTERN_MSC int GMT_get_ellipsoid (struct GMT_CTRL *GMT, char *name);
